move assertUnreachable to shared package
