+++ packages/shared/src/index.ts
@@ -18,5 +18,6 @@ export * from './toSignableCancelOffer'
 export * from './toSignableCreateOffer'
 export * from './types'
 export * from './UserDetails'
+export * from './utils'
 export * from './Withdrawal'
 export * from './WithdrawalWithTokenId'
