--- packages/shared/src/utils/assert.ts
@@ -1,3 +1,0 @@
-export function assertUnreachable(_: never): never {
-  throw new Error('There are more values to handle.')
-}
